Move "utility" functions to more specific files
Deck validation in /deck, customId encoding in interaction event
