+++ src/slash/deck.ts
@@ -1,15 +1,22 @@
-import { RESTPostAPIApplicationCommandsJSONBody } from "discord-api-types/v10";
+import { ButtonStyle, RESTPostAPIApplicationCommandsJSONBody, TextInputStyle } from "discord-api-types/v10";
 import {
+	ActionRowBuilder,
 	AutocompleteInteraction,
+	ButtonBuilder,
+	ButtonInteraction,
 	CacheType,
 	ChatInputCommandInteraction,
+	ModalBuilder,
+	ModalMessageModalSubmitInteraction,
 	SlashCommandBuilder,
+	TextInputBuilder,
 	userMention
 } from "discord.js";
-import { ManualTournament } from "../database/orm";
-import { AutocompletableCommand } from "../SlashCommand";
+import { ManualDeckSubmission, ManualTournament } from "../database/orm";
+import { encodeCustomId } from "../events/interaction";
+import { AutocompletableCommand, ButtonClickHandler, MessageModalSubmitHandler } from "../SlashCommand";
 import { getLogger, Logger } from "../util/logger";
-import { authenticateHost, autocompleteTournament, generateDeckValidateButtons, tournamentOption } from "./database";
+import { authenticateHost, autocompleteTournament, tournamentOption } from "./database";
 
 export class DeckCommand extends AutocompletableCommand {
 	#logger = getLogger("command:deck");
@@ -76,5 +83,132 @@ export class DeckCommand extends AutocompletableCommand {
 
 		const row = generateDeckValidateButtons(tournament);
 		await interaction.reply({ content: outMessage, components: [row] });
+	}
+}
+
+export function generateDeckValidateButtons(tournament: ManualTournament): ActionRowBuilder<ButtonBuilder> {
+	const row = new ActionRowBuilder<ButtonBuilder>().addComponents(
+		new ButtonBuilder()
+			.setCustomId(encodeCustomId("accept", tournament.tournamentId))
+			.setLabel("Accept")
+			.setStyle(ButtonStyle.Success),
+		new ButtonBuilder()
+			.setCustomId(encodeCustomId("reject", tournament.tournamentId))
+			.setLabel("Reject")
+			.setStyle(ButtonStyle.Danger)
+	);
+	return row;
+}
+
+export class AcceptButtonHandler implements ButtonClickHandler {
+	readonly buttonIds = ["accept"];
+
+	async click(interaction: ButtonInteraction, ...args: string[]): Promise<void> {
+		const tournamentIdString = args[0];
+		const modal = new ModalBuilder()
+			.setCustomId(encodeCustomId("acceptModal", tournamentIdString))
+			.setTitle("Accept Deck");
+		const deckLabelInput = new TextInputBuilder()
+			.setCustomId("acceptDeckLabel")
+			.setLabel("What is the deck's theme?")
+			.setStyle(TextInputStyle.Short)
+			.setRequired(false);
+		const actionRow = new ActionRowBuilder<TextInputBuilder>().addComponents(deckLabelInput);
+		modal.addComponents(actionRow);
+		await interaction.showModal(modal);
+	}
+}
+
+export class RejectButtonHandler implements ButtonClickHandler {
+	readonly buttonIds = ["reject"];
+
+	async click(interaction: ButtonInteraction, ...args: string[]): Promise<void> {
+		const tournamentIdString = args[0];
+		const modal = new ModalBuilder()
+			.setCustomId(encodeCustomId("rejectModal", tournamentIdString))
+			.setTitle("Reject Deck");
+		const rejectReasonInput = new TextInputBuilder()
+			.setCustomId("rejectReason")
+			.setLabel("Why is the deck illegal?")
+			.setStyle(TextInputStyle.Paragraph)
+			.setRequired(false);
+		const actionRow = new ActionRowBuilder<TextInputBuilder>().addComponents(rejectReasonInput);
+		modal.addComponents(actionRow);
+		await interaction.showModal(modal);
+	}
+}
+
+export class AcceptLabelModal implements MessageModalSubmitHandler {
+	readonly modalIds = ["acceptModal"];
+
+	async submit(interaction: ModalMessageModalSubmitInteraction, ...args: string[]): Promise<void> {
+		const tournamentIdString = args[0];
+		const tournamentId = parseInt(tournamentIdString, 10);
+		const deck = await ManualDeckSubmission.findOneOrFail({
+			where: {
+				discordId: interaction.user.id,
+				tournamentId: tournamentId
+			},
+			relations: ["tournament"]
+		});
+		const tournament = deck.tournament;
+		const player = await interaction.client.users.fetch(deck.discordId);
+
+		let label: string | undefined = interaction.fields.getTextInputValue("acceptDeckLabel");
+		if (label.length === 0) {
+			label = undefined;
+		}
+
+		// set deck as approved and add label
+		// manual query to workaround bug
+		await ManualDeckSubmission.createQueryBuilder()
+			.update()
+			.set({ approved: true, label })
+			.where("tournamentId = :tournamentId AND discordId = :discordId", {
+				tournamentId,
+				discordId: interaction.user.id
+			})
+			.execute();
+		// provide feedback to player
+		await player.send(`Your deck has been accepted by the hosts! You are now registered for ${tournament.name}.`);
+		// TODO: Give player participant role
+		// log success to TO
+		await interaction.reply(
+			`${userMention(player.id)}'s deck for ${tournament.name} has been approved by ${userMention(
+				interaction.user.id
+			)}`
+		);
+	}
+}
+
+export class RejectReasonModal implements MessageModalSubmitHandler {
+	readonly modalIds = ["rejectModal"];
+
+	async submit(interaction: ModalMessageModalSubmitInteraction, ...args: string[]): Promise<void> {
+		const tournamentIdString = args[0];
+		const deck = await ManualDeckSubmission.findOneOrFail({
+			where: {
+				discordId: interaction.user.id,
+				tournamentId: parseInt(tournamentIdString, 10)
+			},
+			relations: ["tournament"]
+		});
+		const tournament = deck.tournament;
+		const player = await interaction.client.users.fetch(deck.discordId);
+		// clear deck submission
+		await deck.remove(); // do we need to update the link on the player's end?
+		// provide feedback to player
+		let message = `Your deck has been rejected by the hosts. Please update your deck and try again.`;
+		const reason = interaction.fields.getTextInputValue("rejectReason");
+		if (reason.length > 0) {
+			message += `\nReason: ${reason}`;
+		}
+		await player.send(message);
+		// log success to TO
+		await interaction.reply(
+			`${userMention(player.id)}'s deck for ${tournament.name} has been rejected by ${userMention(
+				interaction.user.id
+			)}`
+		);
 	}
 }
